Nodal Conduction and Complex operation
this branch includs the function of nodal conduction and the function of the complex operation, and the nodal funciton is based on the method of the book<Dynamics of Rotating Systems> at the page of 184. While the complex operation is based on the STL lib, and the header file <complex>, and I expand the function of multipla of complex matrix with complex vector.